main menu back key functionality
adds back key functionality to main menu (and attaches the script in inventory correctly - see previous commit). back key here is simple in function - it exits the application directly. note: this will not function in the editor, but was tested with a debug.log that has since been commented out - uncomment to test. commented out as it should not be a thing in final build.